[UI] Исправлена анимация появления желаний Resolve #1221
